# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  rest.ErrorMsg:
    properties:
      error:
        example: null request param
        type: string
    type: object
...
  rest.SuccessMsg:
    properties:
      status:
        example: OK
        type: string
    type: object
...
  storage.Author:
    properties:
      email_address:
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      scholar_ship_profile:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
  storage.AuthorResponse:
    properties:
      author_info:
        $ref: '#/definitions/storage.Author'
      basic_info:
        $ref: '#/definitions/storage.Participant'
    type: object
  storage.Participant:
    properties:
      language:
        description: '''ru'', ''en'''
        type: string
      nickname:
        type: string
      role:
        $ref: '#/definitions/storage.ParticipantRole'
      web3Address:
        type: string
    type: object
  storage.ParticipantRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      AdminRole: 5 Admin
      AdvisorRole: 3 Advisor
      AuthorRole: 2 Author
      GuestRole: 0 Guest
      ReaderRole: 1 Reader
      ValidatorRole: 4 Validator
    x-enum-varnames:
    - GuestRole
    - ReaderRole
    - AuthorRole
    - AdvisorRole
    - ValidatorRole
    - AdminRole
...
  storage.Work:
    properties:
      annotation:
        type: string
      content:
        allOf:
        - $ref: '#/definitions/storage.WorkContent'
        description: BODY INFORMATION
      created_at:
        type: string
      id:
        description: BASE INFORMATION
        type: string
      language:
        type: string
      name:
        type: string
      price:
        type: string
      sources:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  storage.WorkContent:
    properties:
      work_data:
        type: string
    type: object
  storage.WorkResponse:
    properties:
      author_info:
        $ref: '#/definitions/storage.AuthorResponse'
      bookmarked:
        type: boolean
      work:
        $ref: '#/definitions/storage.Work'
    type: object
...
info:
  contact: {}
  description: Specification of interaction with the application
  title: SOW library API
  version: "1.0"
paths:
  /add_bookmark/{work_id}:
    get:
      consumes:
      - application/json
      description: Add a new bookmark to work
      parameters:
      - description: work id to add into bookmarks
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Add a new bookmark
      tags:
      - Bookmarks
...
  /bookmarks:
    get:
      consumes:
      - application/json
      description: Get bookmarks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Get bookmarks
      tags:
      - Bookmarks
...
  /publish_work:
    post:
      consumes:
      - application/json
      description: Publish a new work
      parameters:
      - description: Bearer {JWT token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: work and author info
        in: body
        name: Work
        required: true
        schema:
          $ref: '#/definitions/storage.WorkResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.WorkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Publish a new work
      tags:
      - Publish work
  /purchase_work/{work_id}:
    post:
      consumes:
      - application/json
      description: Purchase particular work
      parameters:
      - description: work id to purchase
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Purchase work
      tags:
      - Purchasing works
  /purchased_works:
    get:
      consumes:
      - application/json
      description: Get purchased works
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Purchased works
      tags:
      - Purchasing works
  /remove_bookmark:
    post:
      consumes:
      - application/json
      description: Remove bookmarks
      parameters:
      - description: author web3 address
        in: path
        name: web3_address
        required: true
        type: string
      - description: Bearer {JWT token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Remove bookmarks
      tags:
      - Bookmarks
...
  /works:
    get:
      consumes:
      - application/json
      description: Get all works depends on role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Get all works
      tags:
      - Works
  /works/{work_id}:
    get:
      consumes:
      - application/json
      description: Get work by id
      parameters:
      - description: work id
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Work by id
      tags:
      - Works
  /works/author/{web3_address}:
    get:
      consumes:
      - application/json
      description: Get list author`s works
      parameters:
      - description: author web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: List author`s works
      tags:
      - Works
  /works_by_key_words:
    get:
      consumes:
      - application/json
      description: Search works by particular words
      parameters:
      - description: words to search for
        in: path
        name: key_words
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Get all works by words
      tags:
      - Works
produces:
- application/json
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"